pyproject.toml
[project]
name = "mumbaihacks2025"
version = "0.1.0"
description = "LangGraph server for Mumbai Hacks 2025"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.121.3",
    "fastapi-health",
    "pydantic-settings<3.0.0,>=2.11.0",
    "PyYAML>=6.0.3",
    "python-dotenv",
    "langchain-openai>=0.3.9",
    "langchain-community==0.3.31",
    "langchain-together>=0.1.0",
    "python-multipart>=0.0.20",
    "langgraph~=0.6.0",
    "colorlog>=6.9.0",
    "starlette>=0.50.0,<1.0.0",
    "pymongo>=4.6.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []

